Refactor authorization policies namespace and imports
Updated the `Policies.cs` file to include a new using directive for `DigitalRightsManagement.Infrastructure.Identity` and changed the namespace from `DigitalRightsManagement.Infrastructure.Identity` to `DigitalRightsManagement.Infrastructure.Authorization` for better organization of authorization policies.
