+++ src/DigitalRightsManagement.Api/Endpoints/UserEndpoints.cs
@@ -1,8 +1,8 @@
 ﻿using Ardalis.Result.AspNetCore;
 using DigitalRightsManagement.Application.UserAggregate;
 using DigitalRightsManagement.Domain.UserAggregate;
+using DigitalRightsManagement.Infrastructure.Authorization;
 using DigitalRightsManagement.Infrastructure.Endpoints;
-using DigitalRightsManagement.Infrastructure.Identity;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
